+++ AdhereRViz/NEWS.md
@@ -2,7 +2,7 @@
 
 In sync with `AdhereR` 0.7.1:
 
-- allows the plotting of event intervals and gaps also for sliding windows and per episodes.
+- allows the plotting of event intervals and gaps also for sliding windows and per episodes (controlled from Aesthetics -> Event intervals).
 
 
 # AdhereRViz 0.2